Lotman: fix Windows CI failures (DB handle leaks, disk usage, sqlite busy)
The native lot engine made the lotman packages build and run on Windows CI
for the first time (they were previously linux-only), surfacing 73 failures
in three buckets:

- Handle leaks (50 tests): tests opened temp SQLite databases and never
  closed them. On unix an open file can be unlinked, so t.TempDir cleanup
  succeeded; on Windows the open handle keeps the file locked and RemoveAll
  fails the test. Add core.Manager.Close() and close the DB in each test
  helper's cleanup (core newTestDB, local_cache newCoreTestManager, lotman
  newAdapterTestDB, cshared resetStateForTest, database/utils tests).

- Disk usage (22 tests): InitLotman probes free disk space, which was stubbed
  as unsupported off-unix. Implement getDiskUsage on Windows via
  GetDiskFreeSpaceEx so the lot manager initializes on Windows too; the !unix
  stub now also excludes windows.

- SQLite busy (1 test): TestSQLiteConcurrentWritesNoBusy asserts BEGIN
  IMMEDIATE + busy_timeout serializes writers with zero SQLITE_BUSY. That
  contract holds on the unix platforms Pelican's server DB deploys on, but the
  pure-Go SQLite driver's Windows file locking can still surface a spurious
  SQLITE_BUSY under contention, and the Windows build is client-only (no server
  DB). Restrict the test to unix.
