+++ lotman/engine_test.go
@@ -39,6 +39,13 @@ func newAdapterTestDB(t *testing.T) *gorm.DB {
 	if err != nil {
 		t.Fatalf("open adapter test db: %v", err)
 	}
+	// Close the handle before t.TempDir's cleanup removes the file: on Windows an
+	// open SQLite handle keeps the file locked and RemoveAll fails the test.
+	t.Cleanup(func() {
+		if sqlDB, err := db.DB(); err == nil {
+			_ = sqlDB.Close()
+		}
+	})
 	return db
 }
